Improve helpers, replace "associatedInstance" with workingDirectory arg
