+++ src/Files.App/Helpers/RecycleBinHelpers.cs
@@ -10,14 +10,12 @@ using System.Threading.Tasks;
 using Windows.ApplicationModel.AppService;
 using Windows.Foundation.Collections;
 using Windows.Storage;
-using Microsoft.UI.Xaml.Controls;
-using Files.App.Shell;
 using Vanara.PInvoke;
 using Files.App.Filesystem;
 
 namespace Files.App.Helpers
 {
-    public class RecycleBinHelpers
+    public static class RecycleBinHelpers
     {
         #region Private Members
 
@@ -25,28 +23,26 @@ namespace Files.App.Helpers
 
         private static readonly Regex recycleBinPathRegex = new Regex(@"^[A-Z]:\\\$Recycle\.Bin\\", RegexOptions.IgnoreCase | RegexOptions.CultureInvariant);
 
-        private Task<NamedPipeAsAppServiceConnection> ServiceConnection => AppServiceConnectionHelper.Instance;
-
         #endregion Private Members
 
-        public async Task<List<ShellFileItem>> EnumerateRecycleBin()
+        public static async Task<List<ShellFileItem>> EnumerateRecycleBin()
         {
             return (await Win32Shell.GetShellFolderAsync(CommonPaths.RecycleBinPath, "Enumerate", 0, int.MaxValue)).Enumerate;
         }
 
-        public async Task<bool> IsRecycleBinItem(IStorageItem item)
+        public static async Task<bool> IsRecycleBinItem(IStorageItem item)
         {
             List<ShellFileItem> recycleBinItems = await EnumerateRecycleBin();
-            return recycleBinItems.Any((shellItem) => shellItem.RecyclePath == item.Path);
+            return recycleBinItems.Any((shellItem) => shellItem.RecyclePath.Equals(item.Path));
         }
 
-        public async Task<bool> IsRecycleBinItem(string path)
+        public static async Task<bool> IsRecycleBinItem(string path)
         {
             List<ShellFileItem> recycleBinItems = await EnumerateRecycleBin();
-            return recycleBinItems.Any((shellItem) => shellItem.RecyclePath == path);
+            return recycleBinItems.Any((shellItem) => shellItem.RecyclePath.Equals(path));
         }
 
-        public bool IsPathUnderRecycleBin(string path)
+        public static bool IsPathUnderRecycleBin(string path)
         {
             if (string.IsNullOrEmpty(path))
             {
@@ -55,12 +51,7 @@ namespace Files.App.Helpers
             return recycleBinPathRegex.IsMatch(path);
         }
 
-        public static async Task S_EmptyRecycleBin()
-        {
-            await new RecycleBinHelpers().EmptyRecycleBin();
-        }
-
-        public async Task EmptyRecycleBin()
+        public static async Task EmptyRecycleBin()
         {
             var ConfirmEmptyBinDialog = new ContentDialog()
             {
@@ -71,7 +62,7 @@ namespace Files.App.Helpers
                 DefaultButton = ContentDialogButton.Primary
             };
 
-            ContentDialogResult result = await this.SetContentDialogRoot(ConfirmEmptyBinDialog).ShowAsync();
+            ContentDialogResult result = await UIHelpers.SetContentDialogRoot(ConfirmEmptyBinDialog).ShowAsync();
 
             if (result == ContentDialogResult.Primary)
             {
@@ -79,12 +70,7 @@ namespace Files.App.Helpers
             }
         }
 
-        public static async Task S_RestoreRecycleBin(IShellPage associatedInstance)
-        {
-            await new RecycleBinHelpers().RestoreRecycleBin(associatedInstance);
-        }
-
-        public async Task RestoreRecycleBin(IShellPage associatedInstance)
+        public static async Task RestoreRecycleBin()
         {
             var ConfirmEmptyBinDialog = new ContentDialog()
             {
@@ -95,21 +81,16 @@ namespace Files.App.Helpers
                 DefaultButton = ContentDialogButton.Primary
             };
 
-            ContentDialogResult result = await this.SetContentDialogRoot(ConfirmEmptyBinDialog).ShowAsync();
+            ContentDialogResult result = await UIHelpers.SetContentDialogRoot(ConfirmEmptyBinDialog).ShowAsync();
 
             if (result == ContentDialogResult.Primary)
             {
                 associatedInstance.SlimContentPage.ItemManipulationModel.SelectAllItems();
-                await this.RestoreItem(associatedInstance);
+                await RestoreItemAsync();
             }
         }
 
-        public static async Task S_RestoreSelectionRecycleBin(IShellPage associatedInstance)
-        {
-            await new RecycleBinHelpers().RestoreSelectionRecycleBin(associatedInstance);
-        }
-
-        public async Task RestoreSelectionRecycleBin(IShellPage associatedInstance)
+        public static async Task RestoreSelectionRecycleBin()
         {
             var ConfirmEmptyBinDialog = new ContentDialog()
             {
@@ -120,25 +101,15 @@ namespace Files.App.Helpers
                 DefaultButton = ContentDialogButton.Primary
             };
 
-            ContentDialogResult result = await this.SetContentDialogRoot(ConfirmEmptyBinDialog).ShowAsync();
+            ContentDialogResult result = await UIHelpers.SetContentDialogRoot(ConfirmEmptyBinDialog).ShowAsync();
 
             if (result == ContentDialogResult.Primary)
             {
-                await this.RestoreItem(associatedInstance);
+                await RestoreItemAsync();
             }
         }
 
-        //WINUI3
-        private ContentDialog SetContentDialogRoot(ContentDialog contentDialog)
-        {
-            if (Windows.Foundation.Metadata.ApiInformation.IsApiContractPresent("Windows.Foundation.UniversalApiContract", 8))
-            {
-                contentDialog.XamlRoot = App.Window.Content.XamlRoot;
-            }
-            return contentDialog;
-        }
-
-        public async Task<bool> HasRecycleBin(string path)
+        public static async Task<bool> HasRecycleBin(string path)
         {
             if (string.IsNullOrEmpty(path) || path.StartsWith(@"\\?\", StringComparison.Ordinal))
             {
@@ -161,17 +132,12 @@ namespace Files.App.Helpers
             return false;
         }
 
-        public bool RecycleBinHasItems()
+        public static bool RecycleBinHasItems()
         {
             return Win32Shell.QueryRecycleBin().NumItems > 0;
         }
 
-        public static async Task S_RestoreItem(IShellPage associatedInstance)
-        {
-            await new RecycleBinHelpers().RestoreItem(associatedInstance);
-        }
-
-        private async Task RestoreItem(IShellPage associatedInstance)
+        private static Task RestoreItemAsync()
         {
             var items = associatedInstance.SlimContentPage.SelectedItems.ToList().Where(x => x is RecycleBinItem).Select((item) => new
             {
@@ -180,20 +146,15 @@ namespace Files.App.Helpers
                     item.PrimaryItemAttribute == StorageItemTypes.File ? FilesystemItemType.File : FilesystemItemType.Directory),
                 Dest = ((RecycleBinItem)item).ItemOriginalPath
             });
-            await associatedInstance.FilesystemHelpers.RestoreItemsFromTrashAsync(items.Select(x => x.Source), items.Select(x => x.Dest), true);
+            return FilesystemHelpers.RestoreItemsFromTrashAsync(items.Select(x => x.Source), items.Select(x => x.Dest), true);
         }
-
-        public static async Task S_DeleteItem(IShellPage associatedInstance)
-        {
-            await new RecycleBinHelpers().DeleteItem(associatedInstance);
-        }
-
-        public virtual async Task DeleteItem(IShellPage associatedInstance)
+        
+        public static Task DeleteItemAsync()
         {
             var items = associatedInstance.SlimContentPage.SelectedItems.ToList().Select((item) => StorageHelpers.FromPathAndType(
                 item.ItemPath,
                 item.PrimaryItemAttribute == StorageItemTypes.File ? FilesystemItemType.File : FilesystemItemType.Directory));
-            await associatedInstance.FilesystemHelpers.DeleteItemsAsync(items, true, false, true);
+            return FilesystemHelpers.DeleteItemsAsync(items, true, false, true);
         }
     }
 }
